more improvements
