+++ nlse/pulse.py
@@ -7,13 +7,18 @@ from __future__ import print_function
 import numpy as np
 from scipy.interpolate import interp1d
 from scipy import constants
-from pynlo.util import FFT_t, IFFT_t
 import warnings
 import scipy.ndimage.interpolation
+import scipy.fftpack as fft
+
+def FFT_t(A,ax=0):
+    return fft.ifftshift(fft.ifft(fft.fftshift(A,axes=(ax,)),axis=ax),axes=(ax,))
+def IFFT_t(A,ax=0):
+    return fft.ifftshift(fft.fft(fft.fftshift(A,axes=(ax,)),axis=ax),axes=(ax,)) 
 
 class Pulse:
     """Class which carried all information about the light field. This class 
-       is a base upon which various cases are built (eg analytic pulses,
+       is a base upon which various cases are built, e.g., Gaussian pulses, 
        CW fields, or pulses generated from experimental data.) """
 
     def __init__(self, frep_MHz = None, n = None):
@@ -23,10 +28,12 @@ class Pulse:
                 warnings.warn("frep should be specified in MHz; large value given.")
         if n is not None:
             self.set_NPTS(n)            
+        
         # Constants, moved here so that module runs through Sphynx autodoc when
         # scipy is Mocked out.
         self._c_nmps = constants.value('speed of light in vacuum')*1e9/1e12 # c in nm/ps
         self._c_mks  = constants.value('speed of light in vacuum') # m/s        
+    
     # Private variables:
     # This set is the minimum number required to completely specify the light
     # field. All other representations are derived from them.
@@ -54,10 +61,7 @@ class Pulse:
     _cache_W_Hz                     = None
     
     _not_ready_msg = 'Pulse class is not yet ready -- set center wavelength, time window, and npts.'    
-    def load_consts(self):
-        r""" Load constants, needed after unpickling in some cases """
-        self._c_nmps = constants.value('speed of light in vacuum')*1e9/1e12 # c in nm/ps
-        self._c_mks  = constants.value('speed of light in vacuum') # m/s                
+       
     ####### Private properties    #############################################
     def __get_w0(self):
         r""" Return center angular frequency (THz) """
@@ -712,6 +716,7 @@ class Pulse:
         mask = ampl**2 > intensity_threshold * np.max(ampl)**2
         gdd     = np.poly1d(np.polyfit(self.W_THz[mask], phase[mask], 2))
         self.set_AW( ampl * np.exp(1j*(phase-gdd(self.W_THz))) )
+    
     def remove_time_delay(self, intensity_threshold = 0.05):
 
         spect_w = self.AW
@@ -720,11 +725,13 @@ class Pulse:
         mask = ampl**2 > (intensity_threshold * np.max(ampl)**2)        
         ld     = np.poly1d(np.polyfit(self.W_THz[mask], phase[mask], 1))
         self.set_AW( ampl * np.exp(1j*(phase-ld(self.W_THz))) )
+    
     def add_time_offset(self, offset_ps):
         """Shift field in time domain by offset_ps picoseconds. A positive offset
            moves the pulse forward in time. """
         phase_ramp = np.exp(-1j*self.W_THz*offset_ps)
         self.set_AW(self.AW * phase_ramp)
+    
     def expand_time_window(self, factor_log2, new_pts_loc = "before"):
         r""" Expand the time window by zero padding.
         Parameters
@@ -753,6 +760,7 @@ class Pulse:
                                     np.zeros(pts_after,)) ))            
         else:
             raise ValueError("new_pts_loc must be one of 'before', 'after', 'even'")
+    
     def rotate_spectrum_to_new_center_wl(self, new_center_wl_nm):
         """Change center wavelength of pulse by rotating the electric field in
             the frequency domain. Designed for creating multiple pulses with same
@@ -761,6 +769,7 @@ class Pulse:
         new_center_THz = self._c_nmps/new_center_wl_nm
         rotation = (self.center_frequency_THz-new_center_THz)/self.dF_THz
         self.set_AW(np.roll(self.AW, -1*int(round(rotation))))
+    
     def interpolate_to_new_center_wl(self, new_wavelength_nm):
         r""" Change grids by interpolating the electric field onto a new
         frequency grid, defined by the new center wavelength and the current
@@ -788,6 +797,7 @@ class Pulse:
         AW_new[self.wl_nm < lower_wl_nm] = 0.0
         AW_new[self.wl_nm > upper_wl_nm] = 0.0
         self.set_AW(AW_new)        
+    
     def clone_pulse(self, p):
         '''Copy all parameters of pulse_instance into this one'''
         self.set_NPTS(p.NPTS)
@@ -803,17 +813,20 @@ class Pulse:
              "AT_re" : self.AT.real.tolist(),
              "AT_im" : self.AT.imag.tolist()}
         return d
+    
     def load_pulse_dict(self, d):    
         self.set_NPTS(d["NPTS"])
         self.set_time_window_ps(d["time_window_ps"])
         self.set_center_wavelength_nm(d["center_wavelength_nm"])
         self._frep_MHz = d["frep_MHz"]
         self.set_AT(np.array(d["AT_re"]) + 1j*np.array(d["AT_im"]) )
+    
     def create_cloned_pulse(self):
         '''Create and return new pulse instance identical to this instance.'''
         p = Pulse()
         p.clone_pulse(self)
         return p               
+    
     def create_subset_pulse(self, center_wl_nm, NPTS):
         """ Create new pulse with smaller frequency span, centered at closest 
             grid point to center_wl_nm, with NPTS grid points and
@@ -858,28 +871,6 @@ class Pulse:
             
         """  
         return np.correlate(abs(self.AT)**2, abs(self.AT), mode='same')  
-        
-    def write_frog(self,
-                 fileloc = 'broadened_er_pulse.dat', # default EDFA spectrum
-                 flip_phase = True):
-        """Save pulse in FROG data format. Grid is centered at wavelength
-        center_wavelength (nm), but pulse properties are loaded from data
-        file. If flip_phase is true, all phase is multiplied by -1 [useful
-        for correcting direction of time ambiguity]. time_window (ps) sets 
-        temporal grid size. 
-        
-        power sets the pulse energy:
-        if power_is_epp is True  then the number is pulse energy [J] 
-        if power_is_epp is False then the power is average power [W], and 
-        is multiplied by frep to calculate pulse energy"""       
-        
-        self.fileloc   = fileloc             
-        phase_data = np.unwrap(np.angle(self.get_AW()))
-        inten_data = np.abs(self.get_AW())
-        wavel_data = self.internal_wl_to_nm(self.wl_nm)
-        
-        # Write pulse data file
-        np.savetxt(self.fileloc, np.vstack((wavel_data, inten_data, phase_data)).T) 
     
     def spectrogram(self, gate_type='xfrog', gate_function_width_ps=0.020, time_steps=500):
         """This calculates a spectrogram, essentially the spectrally-resolved cross-correlation of the pulse.
@@ -994,3 +985,256 @@ class Pulse:
         extent = (np.min(DELAYS), np.max(DELAYS), np.min(FREQS), np.max(FREQS))
         
         return DELAYS, FREQS, extent, np.abs(spectrogram)
+        
+        
+############# Derived Pulses ########
+
+class SechPulse(Pulse):
+    def __init__(self, power, T0_ps, center_wavelength_nm,
+                 time_window_ps = 10., frep_MHz = 100., NPTS = 2**10, 
+                 GDD = 0, TOD = 0, chirp2 = 0, chirp3 = 0,
+                 power_is_avg = False):
+        """Generate a squared-hyperbolic secant "sech" pulse 
+                 A(t) = sqrt(P0 [W]) * sech(t/T0 [ps])
+        centered at wavelength center_wavelength_nm (nm).
+        time_window (ps) sets temporal grid size. 
+        
+        Optional GDD and TOD are in ps^2 and ps^3.
+          
+        Note: The full-width-at-half-maximum (FWHM) is given by 
+                 T0_ps * 1.76   
+        """
+        Pulse.__init__(self, frep_MHz = frep_MHz, n = NPTS)
+        # make sure we weren't passed mks units        
+        assert (center_wavelength_nm > 1.0) 
+        assert (time_window_ps > 1.0 )                
+        self.set_center_wavelength_nm(center_wavelength_nm)        
+        self.set_time_window_ps(time_window_ps)
+                        
+        ### Generate pulse
+        if not power_is_avg:
+            # from https://www.rp-photonics.com/sech2_shaped_pulses.html
+            self.set_AT( np.sqrt(power)/np.cosh(self.T_ps/T0_ps) )
+        else:
+            self.set_AT( 1 / np.cosh(self.T_ps/T0_ps) )
+            self.set_AT(self.AT * np.sqrt( power / ( frep_MHz*1.0e6 * self.calc_epp()) ))
+        
+        self.chirp_pulse_W(GDD, TOD)
+        self.chirp_pulse_T(chirp2, chirp3, T0_ps)            
+        
+class GaussianPulse(Pulse):
+    def __init__(self, power, T0_ps, center_wavelength_nm,
+                 time_window_ps = 10., frep_MHz = 100., NPTS = 2**10, 
+                 GDD = 0, TOD = 0, chirp2 = 0, chirp3 = 0,
+                 power_is_avg = False):
+        """Generate Gaussian pulse A(t) = sqrt(peak_power[W]) * 
+            exp( -(t/T0 [ps])^2 / 2 ) centered at wavelength 
+            center_wavelength_nm (nm). time_window (ps) sets temporal grid
+            size. Optional GDD and TOD are in ps^2 and ps^3.
+        
+        Note: For this definition of a Gaussian pulse, T0_ps is the 
+              full-width-at-half-maximum (FWHM) of the pulse.
+        """
+
+        Pulse.__init__(self, frep_MHz = frep_MHz, n = NPTS)
+        # make sure we weren't passed mks units        
+        assert (center_wavelength_nm > 1.0) 
+        assert (time_window_ps > 1.0 )        
+        self.set_center_wavelength_nm(center_wavelength_nm)
+        self.set_time_window_ps(time_window_ps)        
+        
+        GDD = GDD
+        TOD = TOD
+        
+        # from https://www.rp-photonics.com/gaussian_pulses.html
+        self.set_AT( np.sqrt(power) * np.exp(-2.77*0.5*self.T_ps**2/(T0_ps**2)) ) # input field (W^0.5) 
+        if power_is_avg:            
+            self.set_AT(self.AT * np.sqrt( power / ( frep_MHz*1.0e6 * self.calc_epp()) ))
+        self.chirp_pulse_W(GDD, TOD)
+        self.chirp_pulse_T(chirp2, chirp3, T0_ps)
+    
+class SincPulse(Pulse):
+    def __init__(self, power, FWHM_ps, center_wavelength_nm,
+                 time_window_ps = 10., frep_MHz = 100., NPTS = 2**10, 
+                 GDD = 0, TOD = 0, chirp2 = 0, chirp3 = 0,
+                 power_is_avg = False):
+        """Generate sinc pulse A(t) = sqrt(peak_power[W]) * sin(t/T0)/(t/T0)
+        centered at wavelength center_wavelength_nm (nm).
+        The width is given by FWHM_ps, which is the full-width-at-half-maximum 
+        in picoseconds. T0 is equal th FWHM/3.7909885.
+        time_window_ps sets temporal grid size. Optional GDD and TOD are
+        in ps^2 and ps^3."""
+        Pulse.__init__(self, frep_MHz = frep_MHz, n = NPTS)
+        # make sure we weren't passed mks units        
+        assert (center_wavelength_nm > 1.0) 
+        assert (time_window_ps > 1.0 )                
+        self.set_center_wavelength_nm(center_wavelength_nm)        
+        self.set_time_window_ps(time_window_ps)
+
+        T0_ps = FWHM_ps/3.7909885
+        ### Generate pulse
+        if not power_is_avg:
+            # numpy.sinc is sin(pi*x)/(pi*x), so we divide by pi
+            self.set_AT( np.sqrt(power) * np.sinc(self.T_ps/(T0_ps*np.pi)) ) 
+        else:
+            self.set_AT( 1 / np.sinc(np.pi * self.T_ps/(T0_ps*np.pi)) )
+            self.set_AT(self.AT * np.sqrt( power / ( frep_MHz*1.0e6 * self.calc_epp()) ))
+        
+        self.chirp_pulse_W(GDD, TOD)
+        self.chirp_pulse_T(chirp2, chirp3, T0_ps)   
+        
+class FROGPulse(Pulse):
+    def __init__(self, time_window_ps, center_wavelength_nm, power,frep_MHz = 100., NPTS = 2**10,
+                 power_is_avg = False,
+                 fileloc = '',
+                 flip_phase = True):
+        """Generate pulse from FROG data. Grid is centered at wavelength
+        center_wavelength_nm (nm), but pulse properties are loaded from data
+        file. If flip_phase is true, all phase is multiplied by -1 [useful
+        for correcting direction of time ambiguity]. time_window (ps) sets 
+        temporal grid size. 
+        
+        power sets the pulse energy:
+        if power_is_epp is True  then the number is pulse energy [J] 
+        if power_is_epp is False then the power is average power [W], and 
+        is multiplied by frep to calculate pulse energy"""
+        Pulse.__init__(self, frep_MHz = frep_MHz, n = NPTS)
+        try:
+            self.fileloc   = fileloc
+            # make sure we weren't passed mks units
+            assert (center_wavelength_nm > 1.0) 
+            assert (time_window_ps > 1.0 )
+            self.set_time_window_ps(time_window_ps)
+            self.set_center_wavelength_nm(center_wavelength_nm) # reference wavelength (nm)                         
+            
+            # power -> EPP
+            if power_is_avg:
+                power = power / self.frep_mks
+            
+            # Read in retrieved FROG trace
+            frog_data = np.genfromtxt(self.fileloc)
+            wavelengths = frog_data[:,0]# (nm)
+            intensity   = frog_data[:,1]# (arb. units)
+            phase       = frog_data[:,2]# (radians)
+
+            if flip_phase:
+                phase = -1 * phase
+                            
+            pulse_envelope = interp1d(wavelengths, intensity, kind='linear',
+                                      bounds_error=False,fill_value=0)
+            phase_envelope = interp1d(wavelengths, phase, kind='linear', 
+                                      bounds_error=False,fill_value=0)
+                                      
+            gridded_intensity   = pulse_envelope(self.wl_nm)
+            gridded_phase       = phase_envelope(self.wl_nm)
+
+            # Calculate time domain complex electric field A
+            self.set_AW(gridded_intensity*np.exp(1j*gridded_phase))
+            # Calculate normalization factor  to achieve requested 
+            # pulse energy
+            e_scale = np.sqrt(power / self.calc_epp() )
+            self.set_AT(self.AT * e_scale )
+
+        except IOError:
+            print ('File not found.' )
+
+class NoisePulse(Pulse):
+    def __init__(self, center_wavelength_nm, time_window_ps = 10., NPTS = 2**8,
+                 frep_MHz = None):
+        Pulse.__init__(self, n = NPTS, frep_MHz = frep_MHz)
+        self.set_center_wavelength_nm(center_wavelength_nm)
+        self.set_time_window_ps(time_window_ps)        
+        
+        self.set_AW( 1e-30 * np.ones(self.NPTS) * np.exp(1j * 2 * np.pi *
+                 1j * np.random.rand(self.NPTS)))
+                 
+class CWPulse(Pulse):
+    def __init__(self, avg_power, center_wavelength_nm, time_window_ps = 10.0,
+                 NPTS = 2**8,offset_from_center_THz = None):
+        Pulse.__init__(self, n = NPTS)
+        # make sure we weren't passed mks units
+        assert (center_wavelength_nm > 1.0) 
+        assert (time_window_ps > 1.0 )        
+
+        if offset_from_center_THz is None:            
+            self.set_center_wavelength_nm(center_wavelength_nm)
+            self.set_time_window_ps(time_window_ps)        
+         
+            # Set the time domain to be CW, which should give us a delta function in
+            # frequency. Then normalize that delta function (in frequency space) to
+            # the average power. Note that frep does not factor in here.
+            self.set_AT(np.ones(self.NPTS,))
+            self.set_AW(self.AW * np.sqrt(avg_power) / sum(abs(self.AW)) )
+        else:
+            dF = 1.0/time_window_ps
+            n_offset = np.round( offset_from_center_THz/dF)      
+            
+            center_THz = self._c_nmps/center_wavelength_nm -\
+                                    n_offset * dF
+            center_nm = self._c_nmps / center_THz
+            self.set_time_window_ps(time_window_ps)
+        
+            self.set_center_wavelength_nm(center_nm)
+            aws = np.zeros((self.NPTS, ))
+            aws[int(self.NPTS/2.0) + int(n_offset)  ] = 1.0 *np.sqrt(avg_power)
+            self.set_AW(aws)
+        
+    def gen_OSA(self, time_window_ps, center_wavelength_nm, power, 
+                 power_is_epp = False,
+                 fileloc = '',
+                 log = True, rows = 30): # Yb spectrum
+                 
+        """Generate pulse from OSA data. Grid is centered at wavelength
+        center_wavelength_nm (nm), but pulse properties are loaded from data
+        file. time_window (ps) sets temporal grid size. Switch in place for
+        importing log vs. linear data.
+        
+        power sets the pulse energy:
+        if power_is_epp is True  then the number is pulse energy [J] 
+        if power_is_epp is False then the power is average power [W], and 
+        is multiplied by frep to calculate pulse energy"""
+        
+        try:
+            self.fileloc = fileloc
+            
+            self.set_time_window_ps(time_window_ps)
+                  
+            self.center_wl = center_wavelength_nm                 # reference wavelength (nm)             
+            
+            self.w0 = (2. * np.pi * self.c) / self.center_wl # reference angular frequency                
+            
+            self.setup_grids()
+        
+            if not power_is_epp:                
+                power = power / self.frep
+                
+            # Read in OSA data
+            osa_data = np.genfromtxt(self.fileloc, delimiter = ',', skiprows = rows)                
+            
+            wavelengths = osa_data[:,0]# (nm)
+            wavelengths = self.internal_wl_from_nm(wavelengths)
+                    
+            intensity   = osa_data[:,1]# (arb. units)
+
+            if log:
+                intensity = 10.**(intensity / 10.)
+                
+            freq_abs = self.c/wavelengths
+            freq_abs = np.sort(freq_abs)
+            
+            self.freq_rel = freq_abs - self.c / self.center_wl
+            
+            pulse_envelope = interp1d(self.freq_rel, intensity, kind='linear',
+                                      bounds_error = False, fill_value=0)
+                                      
+            self.gridded_intensity = pulse_envelope(self.V / (2*np.pi))
+            
+            # Calculate time domain complex electric field A
+            self.A  = IFFT_t(self.gridded_intensity)
+            # Calculate normalization factor  to achieve requested 
+            # pulse energy
+            e_scale = np.sqrt(power / self.calc_epp() )
+            self.A  = self.A * e_scale
+                
+        except IOError:
+            print ('File not found.')
